Refactored app.ts page
