+++ public/models/loginDetails.js
@@ -1,1 +1,1 @@
-export {};
+export default ;
