--- src/models/loginDetails.ts
@@ -1,4 +1,0 @@
-export default interface LOGINDETAILS {
-  username: string;
-  password: string;
-}
