+++ src/models/post.ts
@@ -1,6 +1,7 @@
-export default interface POST {
+interface POST {
   userId: string;
   id: string;
   title: string;
   body: string;
 }
+export { POST };
