--- src/models/userDetails.ts
@@ -1,5 +1,0 @@
-export default interface USERDETAILS {
-  firstname: string;
-  lastname: string;
-  username: string;
-}
